+++ src/app/funcionario/funcionario.component.ts
@@ -9,7 +9,9 @@ import { Cidade } from '../cidade/cidade';
 import { Estado } from '../estado/estado';
 import { EstadoService } from '../estado/estado.service';
 import { CidadeService } from '../cidade/cidade.service';
-import { LoginService } from '../login/login.service';
+import {LoginService} from '../login/login.service';
+import {Message} from 'primeng/api';
+import {ConfirmationService} from 'primeng/api';
 
 @Component({
   templateUrl: './funcionario.component.html',
@@ -26,13 +28,18 @@ export class FuncionarioComponent implements OnInit {
   estados: Estado[];
   cidadeEdit = new Cidade();
   estadoEdit = new Estado();
+  msgs: Message[] = [];
   
   constructor(private funcionarioService: FuncionarioService
               , private cargoService: CargoService, private setorService: SetorService
               , private estadoService:EstadoService, private cidadeService:CidadeService
-              , private loginService:LoginService
+              , private loginService: LoginService, private confirmationService: ConfirmationService
 ) {
 
+  }
+
+  hasRole(role: string): boolean {
+    return this.loginService.hasRole(role);
   }
 
   ngOnInit(): void {
@@ -40,10 +47,6 @@ export class FuncionarioComponent implements OnInit {
     this.cargoService.findAll().subscribe(e => this.cargos = e);
     this.setorService.findAll().subscribe(e => this.setores = e);
     this.estadoService.findAll().subscribe(e => this.estados = e);
-  }
-
-  hasRole(role: string): boolean {
-    return this.loginService.hasRole(role);
   }
 
   buscaCidades(estado): void{
@@ -65,6 +68,10 @@ export class FuncionarioComponent implements OnInit {
       this.funcionarioEdit = new Funcionario();
       this.findAll();
       this.showDialog = false;
+      this.msgs = [{severity:'sucess', summary:'Confirmado', detail:'Registro salvo com sucesso'}];            
+    },
+     error => {
+      this.msgs = [{severity:'error', summary:'Erro', detail:'Certifique-se de preencher todos os campos.'}];
     });
   }
 
@@ -74,9 +81,27 @@ export class FuncionarioComponent implements OnInit {
   
   }
 
-  remover(funcionario: Funcionario) {
-    this.funcionarioService.delete(funcionario.id).subscribe(() => {
-      this.findAll();
+  confirmDelete(funcionario: Funcionario) {
+    this.confirmationService.confirm({
+        message: 'Essa ação não poderá ser desfeita',
+        header: 'Deseja remover esse registro?',
+       
+        accept: () => {
+            this.funcionarioService.delete(funcionario.id).subscribe(() => {
+            this.findAll();
+            this.msgs = [{severity:'sucess', summary:'Confirmado', detail:'Registro removido com sucesso'}];
+          },
+          error => {
+            this.msgs = [{severity:'error', summary:'Erro', detail:'Este registro nao pode ser removido.'}];
+          });
+        }
     });
-  }
+}
+
+cancelar(){
+this.showDialog = false;
+this.funcionarioService.findAll().subscribe(e => this.funcionarios = e);
+}
+
+ 
 }
